+++ lib/src/native/realm_core.dart
@@ -413,17 +413,21 @@ class _RealmCore {
   }
 
   static void _stateChangeCallback(Object userdata, int state) {
-    final completer = userdata as Completer<SubscriptionSetState>;
-
-    completer.complete(SubscriptionSetState.values[state]);
+    final completer = userdata as CancellableCompleter<SubscriptionSetState>;
+    if (!completer.isCancelled) {
+      completer.complete(SubscriptionSetState.values[state]);
+    }
   }
 
-  Future<SubscriptionSetState> waitForSubscriptionSetStateChange(SubscriptionSet subscriptions, SubscriptionSetState notifyWhen) {
-    final completer = Completer<SubscriptionSetState>();
-    final callback = Pointer.fromFunction<Void Function(Handle, Int32)>(_stateChangeCallback);
-    final userdata = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
-    _realmLib.realm_sync_on_subscription_set_state_change_async(subscriptions.handle._pointer, notifyWhen.index,
-        _realmLib.addresses.realm_dart_sync_on_subscription_state_changed_callback, userdata.cast(), _realmLib.addresses.realm_dart_userdata_async_free);
+  Future<SubscriptionSetState> waitForSubscriptionSetStateChange(SubscriptionSet subscriptions, SubscriptionSetState notifyWhen,
+      [CancellationToken? cancellationToken]) {
+    final completer = CancellableCompleter<SubscriptionSetState>(cancellationToken);
+    if (!completer.isCancelled) {
+      final callback = Pointer.fromFunction<Void Function(Handle, Int32)>(_stateChangeCallback);
+      final userdata = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
+      _realmLib.realm_sync_on_subscription_set_state_change_async(subscriptions.handle._pointer, notifyWhen.index,
+          _realmLib.addresses.realm_dart_sync_on_subscription_state_changed_callback, userdata.cast(), _realmLib.addresses.realm_dart_userdata_async_free);
+    }
     return completer.future;
   }
 
@@ -666,23 +670,26 @@ class _RealmCore {
 
   Future<RealmHandle> openRealmAsync(RealmAsyncOpenTaskHandle handle, CancellationToken? cancellationToken) {
     final completer = CancellableCompleter<RealmHandle>(cancellationToken);
-    final callback =
-        Pointer.fromFunction<Void Function(Handle, Pointer<realm_thread_safe_reference> realm, Pointer<realm_async_error_t> error)>(_openRealmAsyncCallback);
-    final userData = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
-    _realmLib.realm_async_open_task_start(
-      handle._pointer,
-      _realmLib.addresses.realm_dart_async_open_task_callback,
-      userData.cast(),
-      _realmLib.addresses.realm_dart_userdata_async_free,
-    );
-
+    if (!completer.isCancelled) {
+      final callback =
+          Pointer.fromFunction<Void Function(Handle, Pointer<realm_thread_safe_reference> realm, Pointer<realm_async_error_t> error)>(_openRealmAsyncCallback);
+      final userData = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
+      _realmLib.realm_async_open_task_start(
+        handle._pointer,
+        _realmLib.addresses.realm_dart_async_open_task_callback,
+        userData.cast(),
+        _realmLib.addresses.realm_dart_userdata_async_free,
+      );
+    }
     return completer.future;
   }
 
   static void _openRealmAsyncCallback(Object userData, Pointer<realm_thread_safe_reference> realmSafePtr, Pointer<realm_async_error_t> error) {
     return using((Arena arena) {
-      final completer = userData as Completer<RealmHandle>;
-
+      final completer = userData as CancellableCompleter<RealmHandle>;
+      if (completer.isCancelled) {
+        return;
+      }
       if (error != nullptr) {
         final err = arena<realm_error>();
         bool success = _realmLib.realm_get_async_error(error, err);
@@ -1755,12 +1762,11 @@ class _RealmCore {
     await using((arena) async {
       final response_pointer = arena<realm_http_response>();
       final responseRef = response_pointer.ref;
+      final method = _HttpMethod.values[requestMethod];
+
       try {
         // Build request
         late HttpClientRequest request;
-
-        // this throws if requestMethod is unknown _HttpMethod
-        final method = _HttpMethod.values[requestMethod];
 
         switch (method) {
           case _HttpMethod.delete:
@@ -1788,8 +1794,16 @@ class _RealmCore {
           request.add(utf8.encode(body));
         }
 
+        Realm.logger.log(RealmLogLevel.debug, "HTTP Transport: Executing ${method.name} $url");
+
+        final stopwatch = Stopwatch()..start();
+
         // Do the call..
         final response = await request.close();
+
+        stopwatch.stop();
+        Realm.logger.log(RealmLogLevel.debug, "HTTP Transport: Executed ${method.name} $url: ${response.statusCode} in ${stopwatch.elapsedMilliseconds} ms");
+
         final responseBody = await response.fold<List<int>>([], (acc, l) => acc..addAll(l)); // gather response
 
         // Report back to core
@@ -1816,11 +1830,14 @@ class _RealmCore {
         });
 
         responseRef.custom_status_code = _CustomErrorCode.noError.code;
-      } on SocketException catch (_) {
+      } on SocketException catch (socketEx) {
+        Realm.logger.log(RealmLogLevel.warn, "HTTP Transport: SocketException executing ${method.name} $url: $socketEx");
         responseRef.custom_status_code = _CustomErrorCode.timeout.code;
-      } on HttpException catch (_) {
+      } on HttpException catch (httpEx) {
+        Realm.logger.log(RealmLogLevel.warn, "HTTP Transport: HttpException executing ${method.name} $url: $httpEx");
         responseRef.custom_status_code = _CustomErrorCode.unknownHttp.code;
-      } catch (_) {
+      } catch (ex) {
+        Realm.logger.log(RealmLogLevel.error, "HTTP Transport: Exception executing ${method.name} $url: $ex");
         responseRef.custom_status_code = _CustomErrorCode.unknown.code;
       } finally {
         _realmLib.realm_http_transport_complete_request(request_context, response_pointer);
@@ -2320,12 +2337,14 @@ class _RealmCore {
     controller.onConnectionStateChange(ConnectionState.values[oldState], ConnectionState.values[newState]);
   }
 
-  Future<void> sessionWaitForUpload(Session session) {
-    final completer = Completer<void>();
-    final callback = Pointer.fromFunction<Void Function(Handle, Pointer<realm_error_t>)>(_sessionWaitCompletionCallback);
-    final userdata = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
-    _realmLib.realm_sync_session_wait_for_upload_completion(session.handle._pointer, _realmLib.addresses.realm_dart_sync_wait_for_completion_callback,
-        userdata.cast(), _realmLib.addresses.realm_dart_userdata_async_free);
+  Future<void> sessionWaitForUpload(Session session, [CancellationToken? cancellationToken]) {
+    final completer = CancellableCompleter<void>(cancellationToken);
+    if (!completer.isCancelled) {
+      final callback = Pointer.fromFunction<Void Function(Handle, Pointer<realm_error_t>)>(_sessionWaitCompletionCallback);
+      final userdata = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
+      _realmLib.realm_sync_session_wait_for_upload_completion(session.handle._pointer, _realmLib.addresses.realm_dart_sync_wait_for_completion_callback,
+          userdata.cast(), _realmLib.addresses.realm_dart_userdata_async_free);
+    }
     return completer.future;
   }
 
@@ -2341,8 +2360,8 @@ class _RealmCore {
   }
 
   static void _sessionWaitCompletionCallback(Object userdata, Pointer<realm_error_t> errorCode) {
-    final completer = userdata as Completer<void>;
-    if (completer.isCompleted) {
+    final completer = userdata as CancellableCompleter<void>;
+    if (completer.isCancelled) {
       return;
     }
     if (errorCode != nullptr) {